+++ androidApp/src/main/java/org/rhasspy/mobile/android/uiservices/MicrophoneOverlay.kt
@@ -124,6 +124,7 @@ object MicrophoneOverlay : KoinComponent {
 
         val view = getView()
 
+        overlayWindowManager.updateViewLayout(view, mParams.applySettings())
         view.setViewTreeSavedStateRegistryOwner(lifecycleOwner)
         ViewTreeLifecycleOwner.set(view, lifecycleOwner)
